ai/feat: Replace PostgreSQL with SQLite
Migrated database backend from PostgreSQL to SQLite for single-binary
deployment and operational simplicity.

- Replaced pgx with modernc.org/sqlite (pure Go driver)
- Converted migrations to SQLite syntax with STRICT tables
- Changed schema-based naming to prefixes (identity.* → idp_*, wallabago.* → app_*)
- Switched to INTEGER Unix timestamps with _unix suffix
- Updated queries to SQLite parameter syntax (? instead of $1, $2)
- Embedded migrations using golang-migrate/v4 library with //go:embed
- Changed Entry.ID from int32 to int64 to match SQLite AUTOINCREMENT
- Removed PostgreSQL docker-compose services and testcontainers
- Added WALLABAGO_ prefix to environment variables
- Configured SQLite with WAL mode, foreign keys, and 64MB cache

All 11 BDD scenarios (72 steps) pass.

+++ internal/engines/authz.go
@@ -99,11 +99,11 @@ func (e *HierarchicalAuthorizationEngine) CheckPolicy(ctx context.Context, tx *s
 func (e *HierarchicalAuthorizationEngine) lookupResourceOwner(ctx context.Context, tx *sql.Tx, subject policy.Subject, resourceID any) (string, error) {
 	switch subject {
 	case policy.SubjectEntries:
-		entryID, ok := resourceID.(int32)
+		entryID, ok := resourceID.(int64)
 		if !ok {
 			slog.ErrorContext(ctx, "invalid entry ID type",
 				"resourceID", resourceID,
-				"expectedType", "int32")
+				"expectedType", "int64")
 			return "", policy.ErrForbidden
 		}
 		return e.rbacStorage.GetEntryOwnerID(ctx, tx, entryID)
@@ -133,5 +133,5 @@ type RBACStorage interface {
 	GetUserMaxScope(ctx context.Context, tx *sql.Tx, userID string, subject policy.Subject, operation policy.Operation) (policy.Scope, error)
 
 	// Resource owner lookups
-	GetEntryOwnerID(ctx context.Context, tx *sql.Tx, entryID int32) (string, error)
+	GetEntryOwnerID(ctx context.Context, tx *sql.Tx, entryID int64) (string, error)
 }
